Additional Information /Question
####################
The problem we would like you to solve deals with an array of integers. We want to rotate these integers to a specified number of positions to the left.
If for example you have an array of integers [1,2,3,4,5,6,7] and we would like to rotate 2 positions to the left, the solution expected would be [3,4,5,6,7,1,2].

Note: the solution should be able to handle a position value greater than the number of integers in the array. 
If, for example, we would like to rotate [1,2,3,4,5,6,7] 8 positions to left, it would produce the result [2,3,4,5,6,7,1]. 

Negative values for the positions to rotate will throw an error which the code example should handle gracefully.

Input array[] = [1, 2, 3, 4, 5, 6, 7],
d = 2, ( rotation factor by 2 places, Each element will be shifted to the left) 
n =7 ( number of integers in the array)